team_23-main/visualization_app/requirements.txt
plotly==5.15.0
dash==2.18.1
dash_bootstrap_components==1.6.0
dash-html-components==2.0.0
dash-table==5.0.0
loguru==0.7.2
Requests==2.32.3
gunicorn==23.0.0
google-cloud-bigquery
pandas==2.2.1
db-dtypes==1.3.0